Image-Captioning/playground_requirements.txt
safetensors
transformers==4.28.0
optimum
onnxruntime
onnx
pillow
salesforce-lavis
fastapi
uvicorn
python-multipart